Move shortrange code
