--- src/wr_logic_ai/src/wr_logic_ai/shortrange/shortrange_state_machine.puml
@@ -1,10 +1,0 @@
-@startuml ShortrangeStateMachine
-hide empty description
-
-[*] -right-> shortrange_callback : ShortRangeGoal
-shortrange_callback -right-> VisionNavigation : Vision
-shortrange_callback -up-> success : GPS only
-VisionNavigation -> success : Success
-VisionNavigation -> fail : Fail
-
-@enduml
