+++ src/wr_logic_shortrange/config/vision_params.yaml
