+++ src/wr_logic_shortrange/src/wr_logic_shortrange/vision_navigation.py
@@ -13,10 +13,10 @@ from typing import Tuple, Optional
 import rospy
 from geometry_msgs.msg import PoseStamped
 
-from shortrange_util import ShortrangeStateEnum, ShortrangeState, TargetCache
+from wr_logic_shortrange.shortrange_util import ShortrangeStateEnum, ShortrangeState, TargetCache
 from wr_logic_shortrange.msg import VisionTarget
 from wr_drive_msgs.msg import DriveTrainCmd
-from wr_logic_ai.shortrange.shortrange_util import ShortrangeStateEnum
+from wr_logic_shortrange.vision_navigation import ShortrangeStateEnum
 
 ## Distance from target to stop at (in meters)
 STOP_DISTANCE_M = 3
